decimal numbers in the estimation and spent units
modification to work with decimal numbers in the estimation and spent
units

+++ card.js
@@ -20,14 +20,14 @@ var Card = {
 		if (splits.length >= 2) {
 			var splits2 = splits[0].split(this.secondarySeparator);
 			if (splits2.length == 2) {
-				var value = parseInt(splits2[1]);
+				var value = parseFloat(splits2[1]);
 				if (isNaN(value)) {
 					return 0;
 				}
 				return value;
 			} else {
-				var value = parseInt(splits[0]);
-				if (isNaN(value)) {
+				var value = parseFloat(splits[0]);
+                if (isNaN(value)) {
 					return 0;
 				}
 				return value; 
@@ -39,13 +39,16 @@ var Card = {
 		var trimmedTitle = $.trim(title);
 		if (!stringStartsWith(trimmedTitle, this.startSeparator)){
 			return 0;
-		} 
+		}
+
 		trimmedTitle = trimmedTitle.substring(1, trimmedTitle.length);
 		var splits = trimmedTitle.split(this.mainSeparator);
 		if (splits.length == 2) {
 			var splits2 = splits[0].split(this.secondarySeparator);
 			if (splits2.length == 2) {
-				var value = parseInt(splits2[0]);
+
+				var value = parseFloat(splits2[0]);
+
 				if (isNaN(value)) {
 					return 0;
 				}
